# Dyno: enable PoI instantiation cache and make other perf. optimizations