--- a/frontend/include/chpl/resolution/ResolutionContext.h
+++ b/frontend/include/chpl/resolution/ResolutionContext.h
@@ -500,7 +500,8 @@ class ResolutionContext::GlobalQuery {
   void inactiveStore(RetByVal x) {
     context_->querySetterUpdateResult(QUERY, ap_, std::move(x),
                                       name_,
-                                      isInput_);
+                                      isInput_,
+                                      /* markExternallySet */ false);
   }
 
   bool isRunning() {
@@ -674,7 +675,7 @@ ResolutionContext createDummyRC(Context* context);
 #define CHPL_RESOLUTION_REF_TO_CURRENT_QUERY_HANDLE() (rcquery__)
 
 /** Use this to store a result for any 'CHPL_RESOLUTION_QUERY...' query. */
-#define CHPL_RESOLUTION_QUERY_STORE_RESULT(fn__, rc__, x__, ...) do { \
+#define CHPL_RESOLUTION_QUERY_UNSAFE_STORE_RESULT(fn__, rc__, x__, ...) do { \
     auto rcq__ = rc__->createQueryClass<fn__>(#fn__, __VA_ARGS__); \
     rcq__.inactiveStore(std::move(x__)); \
   } while (0)